requirements.txt
streamlit==1.29.0
pandas==2.1.4
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
Pillow==10.1.0